requirements.txt
django==2.2.9
django-extensions==2.2.9
django-import-export==2.1.0